Imported Upstream version 4.1.9

+++ modules/ims_isc/mark.c
@@ -39,7 +39,7 @@
  *
  * You should have received a copy of the GNU General Public License 
  * along with this program; if not, write to the Free Software 
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  * 
  */
 
@@ -313,7 +313,7 @@ inline int isc_mark_write_route(struct sip_msg *msg, str *as, str *iscmark) {
  */
 int isc_mark_write_psu(struct sip_msg *msg, isc_mark *mark) {
     struct lump *l = msg->add_rm;
-    size_t hlen;
+    int hlen;
     char * hstr = NULL;
     const str *regstate, *sescase;
